# api-docs.yaml
openapi: 3.0.1
info:
  title: Org Job Pricing Engine API
  description: API for calculating pricing and managing sample data, interacting
    with Salesforce via AppLink.
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Local development server
tags:
  - name: Pricing Engine
    description: Quote generation endpoints
  - name: Sample Data
    description: Sample data management endpoints
paths:
  /api/executebatch:
    post:
      operationId: executeBatch
      summary: Submit Batch Pricing Job
      tags:
        - Pricing Engine
      description: Calculate pricing and generate quotes from Opportunities queried
        using the SOQL WHERE clause.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchExecutionRequest"
      responses:
        "202":
          description: Job accepted for processing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchExecutionResponse"
  /api/data/delete:
    post:
      operationId: datadelete
      summary: Submit Sample Data Deletion Job
      tags:
        - Sample Data
      description: Starts a job to delete generate Quotes
      responses:
        "202":
          description: Data operation job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataJobResponse"
  /api/data/create:
    post:
      operationId: datacreate
      summary: Submit Sample Data Creation Job
      tags:
        - Sample Data
      description: Starts a job to create a large amount of Opportunity records.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: []
              properties:
                count:
                  type: integer
                  minimum: 1
                  default: 10
                  description: Number of sample Opportunity records to create (defaults to 10)
      responses:
        "202":
          description: Data operation job accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataJobResponse"
components:
  schemas:
    BatchExecutionRequest:
      type: object
      required:
        - soqlWhereClause
      description: Request to execute a batch process, includes a SOQL WHERE clause to
        extract product information
      properties:
        soqlWhereClause:
          type: string
          description: A SOQL WHERE clause for filtering opportunities
    BatchExecutionResponse:
      type: object
      description: Response includes the unique job ID processing the batch execution
        request.
      properties:
        jobId:
          type: string
          format: uuid
          description: Unique job ID for tracking the worker process
    DataJobResponse:
      type: object
      description: Response includes the unique job ID processing the data operation
        request.
      properties:
        jobId:
          type: string
          format: uuid
          description: Unique job ID for tracking the worker process